Update edit flow for lossless
