+++ packages/web/src/components/data-entry/FormTile.jsx
@@ -26,7 +26,7 @@ import TextArea from 'components/data-entry/TextArea'
 import Dropdown from 'components/navigation/Dropdown'
 import ConnectedRemixSettingsModal from 'components/remix-settings-modal/ConnectedRemixSettingsModal'
 import { RemixSettingsModalTrigger } from 'components/remix-settings-modal/RemixSettingsModalTrigger'
-import SourceFilesModal from 'components/source-files-modal/SourceFilesModal'
+import { StemFilesModal } from 'components/stem-files-modal/StemFilesModal'
 import Switch from 'components/switch/Switch'
 import PreviewButton from 'components/upload/PreviewButton'
 import UploadArtwork from 'components/upload/UploadArtwork'
@@ -256,7 +256,11 @@ const BasicForm = (props) => {
 
   const renderSourceFilesModal = () => {
     return (
-      <SourceFilesModal
+      <StemFilesModal
+        isOriginalAvailable={props.defaultFields.is_original_available}
+        onUpdateIsOriginalAvailable={(value) =>
+          props.onChangeField('is_original_available', value)
+        }
         downloadSettings={props.defaultFields.download}
         onUpdateDownloadSettings={(settings) =>
           props.onChangeField('download', settings)
