+++ packages/web/src/pages/upload-page/fields/StemFilesView.module.css
@@ -1,4 +1,4 @@
-.sourceFilesContainer {
+.stemFilesContainer {
   width: 100%;
   display: flex;
   flex-direction: column;
